Initial separation of code in cloudfoundry input for v1
